Rename the 'crypt_filter' enumerated type to make it less likely to collide with any future definitions (IDENTITY seemed too common).
Add initial support for 'Identity' Crypt Filters. One of the failing
test files used these, its the only test case I've found so far that
does. Now that we have a test file, support it!

Essentially if the Crypt Filter is 'Identity' then we don't use
decryption. We need to be able to handle StmF and StrF separately and,
in the long run, we need to be able to handle dictionaries which
contain Crypt Filters of their own, not just the entries in the /Encrypt
dictionary. Still looking for a test case for these though.