Don't treat missing file /ID as an error in decryption
This really should be an error, but Acrobat and Ghostscript can open
the file by simply treating the missing string as an empty string.

Also, test the empty string as an owner password as well as a user
password. Unbelievably we have a test file where the user password is
not empty, but the owner password is...... Acrobat refuses to open this
file but Ghostscript happily renders it.