Some improvements to the repair code
The code now attempts to spot the trailer dictionary and store it in
ctx->Trailer. To do this we break up the pdfi_read_object into two
parts. The first one reads the object number and 'obj' keyword then
calls pdfi_read_bare_object which only reads an actual PDF object.

We read the trailer dictionary following the 'trailer'
keyword using pdfi_read_bare_object (because uniquely in PDF this
object has no object number).

If we don't already have a trailer dictionary, we use this one.

If we do have a trailer dictionary then we check to see if the new
dictionary has a /Root key. If it does we use the new one, otherwise
we continue with the old one.

This covers the cases of a Linearized file and an incrementally updated
file.

More fixes for the case of a missing endobj. As with earlier commits
we note the offsets of the last 3 objects we read, in case we have no
end obj but do have an 'obj'. In which case we want to rewind the
underlying file and read the object. If the seek fails we just ignore
it (it will if we aren't repairing the file), its no worse than before
and might still be possible to work with.