Optimize spot-checking for ridiculous cases
(From Ken's working branch plus Nancy's changes)

This adds an optimization that was in gs, to basically handle the
file Bug697655.pdf which otherwise takes hours to check.

Create a 'tracking' structure, put the spot dict and transparency flags
in there to reduce the number of parameters being passed.

Add a new CheckedResources array. When we check a resource, mark it off
here and when we come to check an array anew, don't check for spots
or transparency if we've already checked this resource.

Also:
Move checking for a spot-capable device to file open rather than
checking on every page, do it up front.

Rename pdfi_check_Resources_for_transparency() to pdfi_check_Resources()
because the name was misleading anyway.

Various cleanups and comments.