PDF interpreter Ignore non-array Annots
Yet another customer file, so no bug report.

The supplied file is badly broken in several different ways. Most of
these Ghostscript already coped with but the file contains a new fault
we haven't seen before. The same object is referenced by different pages
once as a Font dictionary and once as an Annots array.

Since the object is actually a font dictionary, the page referencing it
as an array fails with an error, because it has the wrong type.

This commit simply ignores such errors and carries on, emitting the
usual warning. Its a little involved as there are several places where
we need to check the type of the object.