Find JPX image even if no Filter specified
If an image has no ColorSpace and isn't an ImageMask, then do a
speculative check for a JPX header and treat as JPX if found.

This fixes tests_private/comparefiles/Bug693625.pdf

Use DefaultRGB for image with no ColorSpace.  It turns out this isn't
actually needed for the test sample, but added it since it is in the
gs version.

Cleaned up some code.