Lots of tidying up surrounding the decryption code
Discard an unused variable in the context (PDFPassword)

Add a PasswordLen to the context. Currently this is set using strlen
but one day we might need to accomadate NULL characters in the password
so this is for future proofing.

Remove the pdfi_apply_SHA256_filter function because we now do the
hashing in place in the pdf_sec.c file. But keep the code for now in
case we might need it for something else.

Fix a memory leak when we encounter an encrypted file with a stream
which has a length of 0 or, theoretically, has a CryptFilter with an
Identity StmF (we don't handle this yet, but we may eventually need to)

Sort out the (non-working) locale_to_utf8 code when we have lilbidn
support. This is still untested code but what was there previously
was just wrong.

Stick our usual #ifndef guards around pdf_file.h which was missing them.

Name change pdfi_read_Encryption to pdfi_initialise_Encryption

Remove the 'R' (Revision) argument from the password checking routines
for revision 5 and 6; we *know* the revision for these, we don't need to
pass it to these routines.

Change the SHA256 hashing code to use the hash functions directly rather
than building streams. Should be faster and more efficient, its
certainly easier to follow....

Add/fix a few comments.

Simplify some code for freeing objects in check_owner_password_R6

Add routines to check passwords which do all the juggling of user,
owner, empty passwords to simplify the code in the initialisation
routine.

Split up the code to read (and validate) the Encryption dictionary
contents from the rest of the initialisation, just to simplify matters.

Add some gs_note_error macros where needed.

Add a 'NONE' to the crypt_type enumerated type so we can use it to tell
that we haven't set up encryption filters yet (differentiates it from
IDENTITY).

Use the CRYPT_NONE enumerated value to simplify the way we detect the
default StdCF entries if they are present, and set up the default
filters if they are not. Remove some now unused variables.

NULL a few pointers when they are freed, just to make sure we don't try
to free them again.