Build changes: allow pdfi to be built-into and called from gs
This meant adding a end_page function pointer to the pdfi context, which should
be set by the client - the "normal" implementation is pdf_impl_end_page() in
pdftop.c, and there is a  dummy version in psi.

To consider: we might want to add the end_page as a paramter passed to
pdfi_create_context().

Plus a *very* noddy integration that runs a PDF in one go from a custom
Postcript operator: .dopdffile (don't forget running from that needs
"--permit-file-read=path/to/my/pdf" command line parameter.