Add a new function to take an existing stream as input
Needed for PostScript integration. If GS is given input on stdin, and
determines the input is PDF, then it writes the input to a temporary
file, and uses that to interpret the PDF. We can't open a temporary
file, so we have to use the existing stream, which we couldn't
previously do.

So add a new function which is handed the input stream, rejig the
existing pdfi_open_file so that it opens a file and passes the stream
to that function.

Finally, do some cleanup refactoring.