Spot overprint in RGB Blend color spaces
If the blending color space is RGB based,
special consideration must be made for doing
overprint with separation and devicen color spaces.

In particular if we are going out to a separation
device we could have RGB+spot support but not
if the separation color was a process name like
Cyan.

Also implement the missing gxdso_pdf14_sep_device
special operation.