Fix 702400: Indeterminate SEGV due to GC confusion.
The GC 'clump_locate' macro relied on PTR_BETWEEN (yes, also a macro)
but on Windows MSC_VER caused pointers to be compared using only the
low 32-bits (unsigned long is 32-bits on Windows, even 64-bit).

Depending on addresses, and when the GC ran, this could result in the
GC logic being confused about addresses. This change works with 32-bit
Windows as well as 64-bit (and may not work with TURBOC on an OLD
CPU with segment+offset addresses, but so what?).