Improve font copying for new PDF interpeter with pdfwrite
Since this commit:

http://git.ghostscript.com/?p=ghostpdl.git;a=commitdiff;h=12c071547d

The 'copied font' stored by pdfwrite has maintained a pointer to the
interpreter's 'dir' (FontDirectory) and used it to mark the names in
the interpreter name table during GC pointer enumeration.

This doesn't work if the interpreter is pdfi and its working inside
the Ghostscript PostScript interpreter, because the enumeration can
take place after the pdfi interpreter has been destroyed. Its also not
necessary in this case, because the pdfi name table can't be relocated.

But for complex reasons (see the lengthy comment in the code) it has a
small but useful benefit when pdfwrite is being driven directly by the
PostScript interpreter.

In order to avoid losing that benefit, we now only store a pointer to
the interpreter's 'dir' member, and only mark the names during
enumeration, if the memory allocator of the font is a garbage-collecting
allocator.