Temporary commit of the 'pdfi in GS' code
Use -dOLDPDF to run the PDF interpreter in PostScript

This command line:
-sOutputFile=/home/ken/out/out.pdf -sDEVICE=pdfwrite -r72 -Z: -sDEFAULTPAPERSIZE=letter -dNOPAUSE -dBATCH -dClusterJob -dJOBSERVER /home/ken/tests/Bug6901014_Arioli-NAG-Warwick.pdf

(obviously adjust paths as required) shows the problem for me under a
debugger and natively (seg fault signal 11).

What I see is that between the 92nd and 93rd call to zPDFDrawPage we
get a call to pdffile_finalize. That is called with cmem = NULL.

Current code ignores that (because previously it would crash there and
then), but when we get to PDFDrawPage for the 93rd time the pdffile
structure's only member, ctx, is no longer valid, it has been
overwritten.

My guess is I'm doing something dumb with the structure and garbage
collector, but I'm scratching my head as to what.

NB the pdffile object should still be on the PostScript operand stack at
this point, because I keep a copy there until we've rendered all the
pages, so we can call .PDFClose on it at the end to free up the
pdfi context and itnerpreter.