lgtm fixes: float * float -> double
When we do float * float, and then use the result in a context
where a double is expected, we implicitly cast the result of the
multiplication to be a double. This means we can overflow
the range of floats during the initial multiplication, when the
result would quite easily fit into a float.

It's better to cast to double to start with, then do the
multiplication (at the cost of potentially being slower - but
not really that much slower these days).

In some cases I've opted NOT to change the operations (merely to
make the casting explicit) so as to minimise diffs.