Free 'DefaultQState' at end of page
Not entirely sure what the 'default QState' is, possibly something
transparency related. Anyway, it was pointing at objects (device
halftone) which had gone away. I think it was probably never valid
to be keeping this, but it shows up with pdfi inside GS because we
meddle with the device on every page.

Freeing it at the end of the page resolves a class of seg faults.