Implement DA string for FreeText annotation
This needs to interpret PDF code from a string.

Add pdfi_interpret_inner_content_string()

Refactor pdfi_interpret_content_stream() to allow calling with
a stream to read content from.

It still wants a "stream_dict" but it looks like it mainly uses this
to lookup resources, so for now I just pass in the Annot dictionary.
It should find the resources in the Page dict anyway.

Also, implement the non-standard 'r' operator that sets the rgb color
from an array.  Add warning message for it.

Also, fix memory leak in pdfi_Tf() in case where the font isn't found.