pdfwrite - rewrite the PSPageOptions code to use non-GC memory
Bug #702370 "PSPageOption injecting unwanted characters"

As Peter rightly comments in the bug this is due to the contents of the
param_string_array not being enumerated when the array pointer is
enumerated.

This seems to be a general problem, but we seem to get away with it for
NeverEmbed and AlwaysEmbed (which are handled in a much more complex
fashion) so I've chosen to leave those alone for now.

For PSPageOptions this commit creates a new gs_param_string_array and
new strings for each of the strings in the param_list. These are
allocated in non-GC memory and the pdfwrite device has complete control
over their lifetime now.

We no longer numerate the member for GC (because it should not be GC'ed)
and we free the array and its contents when we close the device.

This resolves the problem for me.