ps2write - ignore (most) text rendering modes with type 3 fonts
Bug #702233 "ps2 stroked text mitre problem"

The PDF specification says that Tr (text rendering modes) are not
applied to type 3 fonts.

The PDF interpreter normally caters for this by not converting the
glyphs to paths and applying the operation(s) but simply rendering the
text.

However, when the destination is pdfwrite or ps2write, we don't apply
the text rendering mode operations, we simply leave it up to the device
to deal with. Ordinarily this works well, but ps2write had no provision
for the special handling of type 3 fonts, and so was applying the text
rendering mode to those fonts too, resulting in incorrect output.

This commit adds code to the PostScript output of ps2write; this checks
the font type and does not apply the text rendering mode if the font
type is 3. With one exception; if the text rendering mode is 3 (neither
stroke nor fill) then we *do* apply it, because Acrobat does and if we
don't apply it then we draw text that Acrobat does not.