txtwrite - better processing of text in type 3 fonts
Previously we were not able to accumulate the widths of text in a type
3 font. This code leverages (again) off the code in pdfwrite to exit
to the interpreter, run the type 3 CharProc, and pick up the glyph
advance width in setcachedevice.

This gives us better output for type 3 fonts, as we no longer
think that consecutive text is in the same position.