pdfwrite - fix clamping for Acrobat 4 and 5 limits
Bug #702436 "With -dCompatibilityLevel=1.4, gs -sDEVICE=pdfwrite corrupts the contents"

This isn't 'corruption'. Acrobat 4 and 5 have some serious limitations
on the maximum value a real number can take. Since co-ordinates are real
numbers this limits the area which can be addressed on the media.

The example file is very large and, using the default resolution of 720
dpi, some of the content has co-ordinates outside the region which can
be addressed by old versions of Acrobat.

In practice we can see that many parts of our output no longer limit
themselves to these values anyway, so we shjould remove this archaic
limitation.

In the meantime we can work around this specific case by noting that the
clip is rectangular, and will be emitted as a 're' operation, so
clamping all 4 co-ordinates is inappropriate. We need to clamp the llx
and lly co-ordinates, but then its the width and height we need to limit
rather than the urx and ury co-ordinates.