pdfwrite - remove Acrobat 4 & 5 work arounds
Acrobat 4 had a limit of around =/- 16,320 on real numbers (despite the
reference stating it was +/- 32,7670, Acrobat 5 had a limit of 32,767.
These are implementation limits, not architectural limits, and since
co-ordinates are expressed as real numbers they place limits on the
position of objects on the page.

Its clear from Bug #702436 that the clamping used by pdfwrite is not
comprehensive; several paths are emitted that use co-ordinates outside
the permitted range.

Since the implementation is flawed, and the newest of the 2 versions
(Acrobat 5) has been obselete for 15 years, it seems pointless to try
and improve this clamping. Instead this commit removes it entirely.

This removes a number of places where we test values and do some
multiplication, so there should be a (probably immeasurable) performance
gain from this. In addition the accuracy should be (very slightly)
improved because there should be less problems with rounding errors.

This causes a large number of files to demonstrate single pixel shifts
of a few glyphs, due to changes in rounding. In general these are very
minor progressions, though a few could be seen to be equally minor
regressions. Since these mostly (and in the case of regressions,
entirely) occur at low resolution I believe this is acceptable.

One file, Bug694385, when sent to ps2write, exhibits a marked
progression. This is because the clamping wasn't limited to the pdfwrite
device, even though it was inappropriate to limit the ps2write output
this way. A good deal of content is now included, where it was previously
elided.