Refactor downscaler.
Originally the downscaler would just get the bits it wanted from
either get_bits or get_bits_rectangle.

Then we extended the downscaler to integrate ClapTrap, which means
it may have to get its scanlines from that. This was done hackily.

Now, we are pondering integrating scan/skew, which means there will
be a third possible source, and they might be used in different
combinations.

To simplify everything we introduce a "gx_downscaler_liner" mechanism
which wraps up getting a single line of data, and move to using
that everywhere. Adding scan/skew should then plug in using this
much more neatly.