Add CIELAB comparison to bmpcmp
Adding the option -l will perform the difference in CIELAB
color space, provided the application was built with a CMM and
the source files were tiff files that had embedded ICC profiles.

It is necessary to use a threshold > 0 and a window > 1.

The threshold should be interpreted as dE^2 (a Euclidean norm squared
term is used as opposed to the L infinity norm used for the other
color spaces)

As a rule of thumb, dE^2 = 9 is barely noticeable in a side by side
solid color.  Colors in complex images are going to require a larger
threshold.

This adds lcms2mt to the visual studio project.  COLOR_MANAGED