Fixes to spot checking, device params
1) I think Ken's previous fix needs to make sure there is actually a
spot_dict (which means the device can have spots).  Just
unconditionally adding "PageSpotColors" adds it to all devices, even
those without spots.  Possibly harmless, but not right.

2) Fixed a latent memory leak (used to only show up on spot devices,
but now was showing up everywhere, which is why I noticed it).

The call to gs_c_param_list_write() was overwriting the global
"ctx->pdfi_param_list", causing a memory leak on everything contained
in it.  I changed it to use a local param list, which is actually the
same behavior as before, just without the memory leak.

I first tried to just have it do a gs_c_param_list_write_more() to add
to the current list, but this was a massive fail (totally screwing up
other important device parameters like the page size, apparently).

For discussion with Ken: I think because of the way our device can
change depending on transparency and stuff, that global list is not a
great idea.  I noticed just about every place else that uses a device
param list uses a local one just to do what it needs to do.  I think
there is exactly one place where the global variable is used at some
point after it was setup, and coincidentally this is also related to
spot devices (it is in pdfi_set_input_stream).  I wonder if we should
just have that one remaining place query the device directly using a
local param list?  Or possibly just set spot_capable_device in the
same place where we read that global now (and make it no longer a global)?

The global ctx->pdfi_param_list doesn't seem to be gaining much and
just makes things more complex?