Rework display device "DisplayHandle" passing.
Add gsapi_{,de}register_callout API. This allows integrators
to register handlers for "callouts" from gs devices.

The first example of such a callout is the display device.
Previously, this has relied on the gsapi_set_display_callback
API to pass a pointer to a structure into the core, from where
it was hackily poked into the display device structure.

Instead, we now have the display device "callout" to registered
handlers to get the structure and the handle to use.

The legacy API is maintained by the API level implementing
a handler to return the display callback in response to the
devices callout.

The code to do the 'poking' of the display device has therefore
been removed, and replaced by code that checks to see if an
opened device needs reopening after init, if so, opens/closes it.