oss-fuzz 22752: handle broken loca table.
Previously, the code did not differentiate between an empty loca table, a
missing loca table and a broken loca table.

We'll now do so and, in particular, with a broken loca table (in this case,
one with only 1 entry) we'll initialize the single entry to avoid later code
accessing unitialized memory.