Fix CharProc error
It was possible to free a buffer of data (Key) but then try to free it
again if an error occured.

This commit reworks the code to get rid of the buffer in the general
case (we still need it while looking for a .notdef).

Simplifies the code and prevents the seg fault.