Track PostScript file object in pdfi
If the PostScipt code hands us a stream * (rather than us opening a
filename) that stream will be a garbage collected object, which means
it will be relocatable. If the stream (or its contents gets relocated
while the PDF interpreter is using it, then it will suddenly be
referencing the wrong memory.

To solve this we make a new stream in garbage collected but immovable
memory. We then copy the stream we were given into that new stream. We
keep a record of both the streams in the pdffile structure which is
itself a Garbage Collected object.

This means that during GC mark and sweep operations the pdffile object
will correctly inform the GC that ti is referencing both streams (which
caters for the possibility of the user not maintaining a reference to
the original PostScript stream). The pdfi interpreter always uses the
stream structure, so as long as that doesn't move, it doesn't matter
if any of its members do, we will still be able to find them via our
copy of the stream.

At the end we simply copy the stream back to the PostScript stream (so
that any seeks etc will be reflected in it) and zero out the stream
copy we are maintaining.