Handle type 3 fonts that call PDF text operators
We keep a copy of the current text enumerator in the PDF interpreter
context, for the benefit of the d0 and d1 operators, which need it
in order to call gs_text_setcachedevice. When we were done with the
text we set the pointer to NULL.

This is a problem if we run a type 3 font which uses another type 3 font
to draw the text. When we set the enumerator in Tj we are overwriting the
existing enumerator. Worse, on return we set the poitner to NULL, which
would then throw an error back in d0 or d1.

To avoid this, save any existing enumerator, and restore it, around
the text_process() calls.