oss-fuzz 21643: Bounds check a pointer before accessing it.
In this case, immediately after a buffer refill, (as is "normal" for gs stream
buffers) the pointers starts one byte *before* the actual buffer, the state on
reentry meant we'd drop straight to trying dereference the pointer.

This just ensures the pointer gets moved on before we try to use it.