Partial implemenation of AcroForm
This only attempts to render some things that have an AP.
Really need text working to do more here.

Need to use -dShowAcroForm to actual cause this code to run.

Add pdfi_dict_get_ref() because we need to fetch the indirect reference,
not the actual object, for P.  Since P is supposed to refer to to the Page
that contains it, fetching the actual P object gives a cyclical reference in
the cache, which causes a memory leak.  Basically we don't want to resolve
the indirect reference in the 'field' dictionary, we just want to look at it.

I could have used pdfi_dict_get_no_store_R() but it does more than I
need and doesn't have convenient arguments, so I just made this
function instead, which does exactly what is needed with minimal extra
computation.