pdfwrite - don't error out on skipped images
Bug #702471 "eps to pdf OK in 9.50, but failed in 9.51 and 9.52"

Some encode filters, in this case DCTEncode, can throw an error on exit
(premature EOD) if we never write any data to them. This can happen
when pdfwrite sets up to write an image to the output file, but we
never actually write the image, because it is being 'skipped'.

Images can be skipped because they are fully clipped out, unfortunately
we can't detect this before we execute begin_typed_image or we could
avoid a lot of setup.

Anyway, when we get to end_image, if we skipped the image then don't
take any notice of errors when closing the filters used to write the
image data to the PDF file.