Refactor 'clist mutatable devices' from prn ones.
Devices that have to mutate to be clist ones need to be setup
in a particular way. After the standard gx_device header, they
are padded out to a large enough size that the clist header
fields can fit in. Then the device specific fields follow.

gdev_prn devices are the standard example of this. In order
to more easily allow other devices to work in the same way,
we refactor the padding/clist specific fields out into new
macros.