Improve usage of 'get_colorname_string'
This function is supplied to a couple of places, and stored. Its
purpose is to retrieve a string for a colour plane, given an 'index'.
The index is an entry in the interpreter's name table.

The implementations of get_colorname_string take a memory_t pointer
and then use that to get a pointer to the interpreter, which they need
because the interpreter context is where the name table is stored.

The problem with the existing usage is that we were supplying the
device's memory_t pointer. However, if the halftone or colour space had
been created by a different interpreter to the device (which can happen
if we are running the PDF interpreter inside the PostScript interpreter)
then the interpreter context would be incorrect.

We can solve this easily enough by using the colour space or halftone
memory pointer. This is much safer since we can reasonably assume that
the interpreter which created the object is the same one which stored
the colour name in its name table.