Remove clist "is_printer" field.
Currently, the clist has an "is_printer" field, which is used to
tell it how to forward dev_spec_op calls. This seems evil to me;
a base class should not really need to know what class is being
derived from it to know how to behave.

Instead, introduce a function pointer that says where
dev_spec_ops should be forwarded to.