More work on gs_get_colorname_string
Commit 14c70d4e9 was, unfortunately, insufficient for the new PDF
interpreter. The gs_lib_ctx pointer does not give us access back to the
interpreter context, only to the *PostScript* interpreter name table.

So instead we now pass a graphics state pointer to
gs_get_colorname_string(). This is only used for halftones currently,
and even future use seems likely to be limited to objects like colour
spaces, which are of course contained in a graphics state, so this
seems relatviely future proof.

For PostScript the graphics state contains a memory_t pointer which we
can use to get to the PostScript name table. For the PDF interpreter
we already store some state in the 'client_data' of the graphics state
and we will extend that to include a pointer to the interpreter context
which will allow us to get the string from its name table.