Alter pdfi_separation_name_from_index to match new GS code
We modified gs_get_colorname_string to take a graphics state pointer
instead of a memory_t pointer, so that we would be able to fins the
correct interpreter to access the correct name table.

This obviously needs us to modify the code in the PDF itnerpreter as
well.

This introduces two seg faults reading the output from pdfwrite which
I am unable to duplicate locally on Windows or Linux. I'm choosing
to accept those for now and will try to debug them remotely.

It also shows a progression in x_blend_modes.pdf, reversing an earlier
regression we had chosen to accept temporarily. I have no idea why....