When building on arm64 with gcc, disable libpng's neon optimizations.
Bug 702486 - Ghostscript 9.52 does not build on arm64

Ghostscript 9.52 with the libraries which it is including builds
perfectly on i386, amd64, s390x, and armhf, but not on ppc64el
and arm64. Problem are internal optimizations in the libpng
shipped with Ghostscript.

For ppc64el the problem was already solved after the 9.52
release:

http://git.ghostscript.com/?p=ghostpdl.git;a=commitdiff;h=1c9bd8e0aa4
https://bugs.ghostscript.com/show_bug.cgi?id=702236

This fix is for arm64 is Based on this and on a similar fix for
arm64 on another free software project

https://github.com/imagemin/optipng-bin/issues/97

This makes it also build on arm64, but it is not tested whether
it still correctly works on arm64 after the fix.