Fix memory leak in pdfi running under gs
The gs_grestore() expects there to be another saved context around it.
When there's not, it does something strange (does a gsave when you
asked for a grestore -- really?).

It was causing a memory leak because the null device rc got
incremented, such that it never got freed.  Maybe other things were
leaking also, that's just the one I noticed.

I couldn't find a good way to wrap a gsave/grestore around this, so I
changed it to call gs_grestore_only() which doesn't do the extra
gsave nonsense.