Fix PageSize for gs/pdfi
In this case, we get the PageSize from the .MediaSize which gs
has already setup for us.

It wasn't being setup (was [0 0 0 0]), so transparency didn't work.
This is all a bit hacky, but hopefully it's an improvement.
I don't understand whether it will work if the page is rotated.

This should fix some segfaults, plus improve rendering for transparency cases.

Also moved the gs_setblendmode() so it will happen in both cases.
I noticed the value was wrong if we rely on what's coming from gs.