Coverity 359871, 359872: un-init variable and conditional always true
A return code could remain unset in an otherwise successful completion of the
function.

The new FAPI interface switch from using a signed value of -1 to the "standard"
special value of GS_NO_CHAR, and I missed a case of that - since GS_NO_CHAR is
positive, the existing test of >= 0 was flagged as always true.