Improve get_bits_rectangle for planar case.
The implementation of get_bits_rectangle to read planar data from
a planar device, only implements GB_RETURN_POINTER. If this isn't
in the request (i.e. the caller uses GB_RETURN_COPY to get a copy
of the data in its own buffers), then it cannot cope.

Here we extend the routine to try to use a GB_RETURN_POINTER
request internally, and to copy the returned data into the
callers buffers.

Callers are of course already free to do this operation themselves,
but this avoids the needless repetition of the code in every
caller.