Ensure we have two gstates to make the graphics library happy
The graphics library assumes we will always have an initial gstate and
at least one more gstate (so the initial gstate is saved). We were not
doing that.

In addition we need to use gs_grestore_only to clean up, because
gs_grestore will not free the last saved state, if we ever get back to
the point of the initial gstate, gs_grestore immediately does a gsave.
gs_grestore_only doesn't do that, so we use it to clear the stack of
saved states.

This also reverses Nancy's earlier commit to address a memory leak when
running pdfi inside Ghostscipt. Reversing this does reintroduce the
memory leak(s), but these seem to be essentially the same as when
running the same file in pdfi on its own. Obviously we need to address
the leaks but I think this suggests that this was not the correct fix.