Fix Bug 701852. Problems with the plane_extract device.
First, the plane_begin_typed_image function used the gs_gstate_copy_temp,
but since this was allocated as a st_gs_gstate, and the reference counts
were not handled (as the comment describes), freeing that object would
decrement the ref_counts, causing premature free of objects.

Remove this dangerous function.

Change to use gs_gstate_copy so we have a proper gs_gstate structure.

Using gs_gstate_copy from a clist playback gstate requires that we have
client colors and device colors in the color[0] and color[1] elements.
Allocate, and initialize these, then free upon exit from clist_playback.

Also, the mem device underlying the plane_extract device needs to be
retained so that it is not prematurely freed.