Move the device transparency setup into pdfi_check_page()
This way it gets done the same for pdfi and gs/pdfi modes,
and is in the same place as the spots.

This fixes problems with gs/pdfi -- the device needs to be
reinitialized to know it has transparency, so it can figure out its
memory requirements.  This fixes files where the page is big enough
that it needs to use clist for pdf14.  It wasn't being properly
detected so later there was a memory allocation error (which can cause
a segfault).

File: tests_private/comparefiles/Bug692933.pdf